refactor(dev-core): break shared→cli dependency cycle
Move `discoverProject` and the workspace-store fns (getWorkspacePath,
readWorkspace, parseWorkspace, writeWorkspace) down into shared/adapters.
discoverProject now takes a `localPath?` param instead of importing the
cli-only cwd-resolver — the cli wrapper injects `detectLocalPath(repo)` at
the boundary. Collapse the duplicate WorkspaceProject/Workspace/
VercelProjectRef interfaces to the sole definition in ports/workspace.ts.
cli/lib/{github-discovery,workspace-store}.ts become thin re-exports.

git-workspace.ts now has zero cli imports — the dependency arrow points
one way (cli → shared). Behavior-preserving; full suite green.

Slice 4 of #195 (SC#1).
